refactoring dsl_parser into separate files
