+++ src/dsl/src/dsl_parser/dsl_to_dict.py
@@ -1,0 +1,69 @@
+"""This module parses a dsl file and outputs an intermediate stage"""
+
+# for compatibility with python 3.7 and 3.8
+from __future__ import annotations
+
+from lark import Lark, Transformer
+from lark.lexer import Token
+from pathlib import Path
+from warnings import warn
+import json
+
+from dsl_parser import transformers, simple_transformers
+
+
+grammar_path = Path(__file__).parent.parent / "grammars"
+grammar_f = open(grammar_path / "dsl_grammar.lark")
+dsl_grammar = grammar_f.read()
+grammar_f.close()
+
+parser = Lark(dsl_grammar, parser='earley', import_paths=[grammar_path])
+
+
+# main outward-facing function
+def export_dict(file_str, debug=False, debug_modes=[]):
+    """Parses the language and returns an intermediate stage consisting
+    of python dictionaries"""
+    tree = parser.parse(file_str)
+    if debug and "dsl-tree" in debug_modes:
+        print(tree.pretty())
+    intermediate = TreeToDict().transform(tree)
+    if debug and "intermediate" in debug_modes:
+        print(json.dumps(intermediate, indent=2))
+    return intermediate
+
+
+class TreeToDict(Transformer):
+    #for documentation of these transformations, see transformers.py
+    game = transform_game
+
+    room = transform_room
+
+    connections = transform_connections
+
+    item = transform_item
+    
+    action = transform_action
+
+    misplaced_property = transform_misplaced
+
+    #for documentation of the below transformations, see simple_transformers.py
+
+    ESCAPED_STRING = mod_ESCAPED_STRING
+    
+    starg_g = mod_start_g
+
+    end_g = mod_end_g
+
+    id = mod_id_label
+
+    action_ids = mod_action_ids
+
+    location = mod_location
+
+    phrase = mod_phrase
+
+
+    connection = tuple
+    property = tuple
+
